moved parsing logic to duckStringParser.js
