+++ src/library/duckStringParser.js
@@ -1,0 +1,64 @@
+function getObject(input, start, end) {
+   const itemString = input.substring(start, end);
+   return parseInt(itemString, 16); 
+}
+
+function getHat(input) {
+    return getObject(input, 1, 3);
+}
+
+function getEyes(input) {
+    return getObject(input, 3, 5);
+}
+
+function getBeak(input) {
+    return getObject(input, 5, 7);
+}
+
+function getWings(input) {
+    return getObject(input, 7, 9);
+}
+
+function getSmoke(input) {
+    return getObject(input, 9, 11);
+}
+
+function getTail(input) {
+    return getObject(input, 11, 13);
+}
+
+function getItem(input) {
+    return getObject(input, 13, 15);
+}
+
+function getBeakColor(input) {
+    return getObject(input, 15, 16);
+}
+
+function getFeatherColor(input) {
+    return getObject(input, 16, 17);
+}
+
+function getData(input) {
+    const hat = getHat(input);
+    const eyes = getEyes(input);
+    const beak = getBeak(input);
+    const wings = getWings(input);
+    const smoke = getSmoke(input);
+    const tail = getTail(input);
+    const item = getItem(input);
+    const beakColor = getBeakColor(input);
+    const featherColor = getFeatherColor(input);
+
+    return { hat,
+            eyes,
+            beak,
+            wings,
+            smoke,
+            tail,
+            item,
+            beakColor,
+            featherColor};
+}
+
+module.exports = { getData };
